Endgame: One generator to rule them all
